A - move all in misc to easly found them
