+++ app/src/main/java/com/poupa/vinylmusicplayer/misc/queue/DynamicElement/AlbumShuffling/AlbumShufflingUtil.java
@@ -1,4 +1,4 @@
-package com.poupa.vinylmusicplayer.util.DynamicElement;
+package com.poupa.vinylmusicplayer.misc.queue.DynamicElement.AlbumShuffling;
 
 import java.lang.reflect.Type;
 import java.util.ArrayList;
@@ -11,7 +11,7 @@ import com.google.gson.Gson;
 import com.google.gson.JsonSyntaxException;
 import com.google.gson.reflect.TypeToken;
 import com.poupa.vinylmusicplayer.App;
-import com.poupa.vinylmusicplayer.model.AlbumShufflingCriteria;
+
 
 public final class AlbumShufflingUtil {
