+++ app/src/main/java/com/poupa/vinylmusicplayer/misc/queue/DynamicElement/DynamicElementPreferenceFragment.java
@@ -1,0 +1,8 @@
+package com.poupa.vinylmusicplayer.misc.queue.DynamicElement;
+
+import androidx.fragment.app.Fragment;
+
+public abstract class DynamicElementPreferenceFragment extends Fragment {
+    public abstract void reset();
+    public abstract void ok();
+}
