--- app/src/main/java/com/poupa/vinylmusicplayer/ui/fragments/misc/DynamicElementPreferenceFragment.java
@@ -1,8 +1,0 @@
-package com.poupa.vinylmusicplayer.ui.fragments.misc;
-
-import androidx.fragment.app.Fragment;
-
-public abstract class DynamicElementPreferenceFragment extends Fragment {
-    abstract void reset();
-    abstract void ok();
-}
